Crate apns2

A library for sending push notifications to iOS devices using Apple's APNS API. Supports certificate based authentication through apns2::client::CertificateClient and JWT token based authentication through apns2::client::TokenClient.

If using JWT tokens for authentication, apns2::apns_token::ApnsToken can be used for generating and holding tokens, allowing re-use and renewal.

The apns::client::ProviderResponse does not block until using the recv_timeout. The request is handled in another thread and the response is sent through a channel to the thread calling the method.

Modules

apns_token
A module for APNS JWT token management.

client
The APNs connection handling modules. TokenClient for connections using JWT authentication, CertificateClient when using a sertificate and a private key to authenticate. ProviderResponse handles responses and maps the results to APNSStatus and APNSError.

notification
A helper struct for generating an APNS request.

payload