Struct apns2::client::CertificateClient

pub struct CertificateClient {
    pub client: Client,
}

Fields
client: Client
Methods

impl CertificateClient

Creates a new connection to APNs using a certificate and private key to an application. The connection is only valid for one application.

The response for push is asynchorous for better throughput.

Example

use apns2::client::CertificateClient;
use apns2::payload::{Payload, APSAlert};
use apns2::notification::{Notification, NotificationOptions};
use std::fs::File;
use std::time::Duration;

// Can be anything that implements the `Read` trait.
let mut cert_file = File::open("/path/to/certificate.pem").unwrap();
let mut key_file = File::open("/path/to/key.pem").unwrap();

let client = CertificateClient::new(false, &mut cert_file, &mut key_file).unwrap();
let alert = APSAlert::Plain(String::from("Hi there!"));
let payload = Payload::new(alert, "default", Some(1u32), None, None);
let options = NotificationOptions { ..Default::default() };
let request = client.push(Notification::new(payload, "apple_device_token", options));

// Block here to get the response.
let response = request.recv_timeout(Duration::from_millis(2000));

println!("{:?}", response);

fn new<'a, R: Read>(sandbox: bool,
certificate: &mut R,
private_key: &mut R)
-> Result<CertificateClient, ProviderError>

Create a new connection to APNs with custom certificate and key. Can be used to send notification to only one app.

fn push<'a>(&self, notification: Notification) -> ProviderResponse

Send a notification.