Struct apns2::client::TokenClient

pub struct TokenClient {
    pub client: Client,
}

Fields
client: Client
Methods

impl TokenClient

Creates a new connection to APNs using the system certificates. When sending notifications through this type of connection, one must attach a valid JWT token with every request using either APNSToken or an own implementation. The same connection can be used to send notifications to multiple applications.

The response for push is asynchronous for better throughput.

Examples

use apns2::client::TokenClient;
use apns2::apns_token::APNSToken;
use apns2::payload::{Payload, APSAlert};
use apns2::notification::{Notification, NotificationOptions};
use std::fs::File;
use std::time::Duration;

// Can be anything that implements the `Read` trait.
let der_file = File::open("/path/to/key.der").unwrap();

let apns_token = APNSToken::new(der_file, "TEAMID1234", "KEYID12345").unwrap();

// Example certificate file from Ubuntu
let client = TokenClient::new(false, "/etc/ssl/certs/ca-certificates.crt").unwrap();

let alert = APSAlert::Plain(String::from("Hi there!"));
let payload = Payload::new(alert, "default", Some(1u32), None, None);
let options = NotificationOptions { ..Default::default() };
let request = client.push(Notification::new(payload, "Hi there!", options), apns_token.signature());

// Block here to get the response.
let response = request.recv_timeout(Duration::from_millis(2000));

println!("{:?}", response);

fn new<'a>(sandbox: bool,
certificates: &str)
-> Result<TokenClient, ProviderError>

Create a new connection to APNs. certificates should point to system ca certificate file. In Ubuntu it's usually /etc/ssl/certs/ca-certificates.crt.

fn push(&self, notification: Notification, apns_token: &str) -> ProviderResponse

Send a push notification with a JWT signature.