Struct apns2::payload::APS

pub struct APS {
    pub alert: Option<APSAlert>,
    pub badge: Option<u32>,
    pub sound: Option<String>,
    pub content_available: Option<u32>,
    pub category: Option<String>,
}

The APS can contain one or more properties that specify the following user notification types: - an alert message to display to the user - a number to badge the app icon with - a sound to play

Fields
alert: Option<APSAlert>

If this property is included, the system displays a standard alert or a banner, based on the user’s setting.

badge: Option<u32>

The number to display as the badge of the app icon.

sound: Option<String>

The name of a sound file in the app bundle or in the Library/Sounds folder of the app’s data container.

content_available: Option<u32>

Provide this key with a value of 1 to indicate that new content is available.

category: Option<String>

Provide this key with a string value that represents the identifier property.