Struct apns2::payload::APSLocalizedAlert

pub struct APSLocalizedAlert {
    pub title: String,
    pub body: String,
    pub title_loc_key: Option<String>,
    pub title_loc_args: Option<Vec<String>>,
    pub action_loc_key: Option<String>,
    pub loc_key: Option<String>,
    pub loc_args: Option<Vec<String>>,
    pub launch_image: Option<String>,
}

Child properties of the alert property.

Fields
title: String

A short string describing the purpose of the notification

body: String

The text of the alert message.

title_loc_key: Option<String>

The key to a title string in the Localizable.strings file for the current localization.

title_loc_args: Option<Vec<String>>

Variable string values to appear in place of the format specifiers in title-loc-key.

action_loc_key: Option<String>

If a string is specified, the system displays an alert that includes the Close and View buttons.

loc_key: Option<String>

A key to an alert-message string in a Localizable.strings file for the current localization.

loc_args: Option<Vec<String>>

Variable string values to appear in place of the format specifiers in loc-key.

launch_image: Option<String>

The filename of an image file in the app bundle. The image is used as the launch image when users tap the action button or move the action slider.